[ALLUXIO-1910] Move ValidateConf to alluxio.cli package
